requirements.txt
alembic==1.0.8
bleach==3.1.0
blinker==1.4
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-Images==3.0.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Moment==0.7.0
Flask-PageDown==0.2.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
ForgeryPy3==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
Markdown==3.1
Werkzeug==0.15.2
WTForms==2.2.1
flask-request-validator==2.1.0
Flask-HTTPAuth==3.2.4
geopy==1.19.0
gunicorn==19.9.0
Flask-SSLify==0.1.5
# psycopg2==2.7.3 --no-binary psycopg2
# scikit-learn==0.20.3
# scipy==1.2.1
# numpy==1.16.3